+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/mick/apps/fb-challenge/src/App.js":"1","/home/mick/apps/fb-challenge/src/index.js":"2","/home/mick/apps/fb-challenge/src/Header.jsx":"3","/home/mick/apps/fb-challenge/src/Nav.jsx":"4","/home/mick/apps/fb-challenge/src/Shipping.jsx":"5","/home/mick/apps/fb-challenge/src/Billing.jsx":"6","/home/mick/apps/fb-challenge/src/ReviewOrder.jsx":"7"},{"size":1063,"mtime":1612036154021,"results":"8","hashOfConfig":"9"},{"size":521,"mtime":1612034958244,"results":"10","hashOfConfig":"9"},{"size":206,"mtime":1612033192452,"results":"11","hashOfConfig":"9"},{"size":442,"mtime":1612035937415,"results":"12","hashOfConfig":"9"},{"size":83,"mtime":1612036233024,"results":"13","hashOfConfig":"9"},{"size":80,"mtime":1612035504953,"results":"14","hashOfConfig":"9"},{"size":164,"mtime":1612036268248,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"18"},"16ig0ww",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/home/mick/apps/fb-challenge/src/App.js",["31"],"import { useState, useRef, useEffect } from 'react';\nimport { Route } from 'react-router-dom';\nimport Header from './Header';\nimport Nav from './Nav'\nimport Billing from './Billing';\nimport Shipping from './Shipping'\nimport ReviewOrder from './ReviewOrder';\n\nimport './App.css'\n\nfunction App() {\n  const [headerHeight, setHeaderHeight] = useState(0);\n  const [yAxis, setYAxis] = useState(0);\n  const headerRef = useRef();\n\n  useEffect(() => {\n    setHeaderHeight(headerRef.current.clientHeight);\n    window.addEventListener(\"scroll\", yVal);\n    return () => {\n      window.removeEventListener(\"scroll\", yVal);\n    };\n  });\n\n  const yVal = () => {\n    setYAxis(window.pageYOffset);\n  };\n\n  return (\n    <div className='App'>\n      <Header headerRef={headerRef} />\n      <Nav yAxis={yAxis} headerHeight={headerHeight} />\n      <Route path='/shipping'>\n        <Shipping />\n      </Route>\n      <Route path='/billing'>\n        <Billing />\n      </Route>\n      <Route path='review-order'>\n        <ReviewOrder />\n      </Route>\n    </div>\n  );\n}\n\nexport default App;\n","/home/mick/apps/fb-challenge/src/index.js",[],"/home/mick/apps/fb-challenge/src/Header.jsx",[],"/home/mick/apps/fb-challenge/src/Nav.jsx",[],"/home/mick/apps/fb-challenge/src/Shipping.jsx",[],"/home/mick/apps/fb-challenge/src/Billing.jsx",[],"/home/mick/apps/fb-challenge/src/ReviewOrder.jsx",["32"],{"ruleId":"33","severity":1,"message":"34","line":16,"column":3,"nodeType":"35","endLine":16,"endColumn":12,"suggestions":"36"},{"ruleId":"37","severity":1,"message":"38","line":1,"column":10,"nodeType":"35","messageId":"39","endLine":1,"endColumn":18},"react-hooks/exhaustive-deps","React Hook useEffect contains a call to 'setHeaderHeight'. Without a list of dependencies, this can lead to an infinite chain of updates. To fix this, pass [] as a second argument to the useEffect Hook.","Identifier",["40"],"no-unused-vars","'useState' is defined but never used.","unusedVar",{"desc":"41","fix":"42"},"Add dependencies array: []",{"range":"43","text":"44"},[620,620],", []"]
+[{"/home/mick/apps/fb-challenge/src/App.js":"1","/home/mick/apps/fb-challenge/src/index.js":"2","/home/mick/apps/fb-challenge/src/Shared/Layout.jsx":"3","/home/mick/apps/fb-challenge/src/Components/Nav/Nav.jsx":"4"},{"size":1251,"mtime":1612038932937,"results":"5","hashOfConfig":"6"},{"size":521,"mtime":1612034958244,"results":"7","hashOfConfig":"6"},{"size":133,"mtime":1612038818281,"results":"8","hashOfConfig":"6"},{"size":443,"mtime":1612038901722,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"16ig0ww",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/mick/apps/fb-challenge/src/App.js",["18"],"/home/mick/apps/fb-challenge/src/index.js",[],"/home/mick/apps/fb-challenge/src/Shared/Layout.jsx",[],"/home/mick/apps/fb-challenge/src/Components/Nav/Nav.jsx",[],{"ruleId":"19","severity":1,"message":"20","line":17,"column":3,"nodeType":"21","endLine":17,"endColumn":12,"suggestions":"22"},"react-hooks/exhaustive-deps","React Hook useEffect contains a call to 'setHeaderHeight'. Without a list of dependencies, this can lead to an infinite chain of updates. To fix this, pass [] as a second argument to the useEffect Hook.","Identifier",["23"],{"desc":"24","fix":"25"},"Add dependencies array: []",{"range":"26","text":"27"},[755,755],", []"]
