+++ src/App.css
@@ -1,5 +1,4 @@
 .App {
   max-width: 1600px;
   margin:0 auto;
-  min-height:120vh;
 }
