+++ src/Shared/Layout.css
@@ -1,0 +1,3 @@
+.layout{
+    min-height:120vh;
+}
