+++ src/Shared/Layout.jsx
@@ -1,0 +1,5 @@
+import './Layout.css'
+const Layout = (props) => {
+  return <div className='layout'>{props.children}</div>;
+};
+export default Layout;
